requirements.txt
yt-dlp
opencv-python
webrtcvad-wheels
numpy
PyYAML
colorama